# web/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "HelveticOps.Api",
    "version": "1.0"
  },
  "paths": {
    "/api/work-orders": {
      "get": {
        "tags": [
          "Work orders"
        ],
        "summary": "List and filter work orders",
        "description": "Returns a safely sorted, server-paged work-order projection. Page size is limited to 100.",
        "operationId": "ListWorkOrders",
        "parameters": [
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "Status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/WorkOrderStatus"
            }
          },
          {
            "name": "Priority",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/WorkOrderPriority"
            }
          },
          {
            "name": "Site",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SlaRisk",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Sort",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/WorkOrderSort"
            }
          },
          {
            "name": "Direction",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SortDirection"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderListItemPagedResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Work orders"
        ],
        "summary": "Create a work order",
        "operationId": "CreateWorkOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/work-orders/{id}": {
      "get": {
        "tags": [
          "Work orders"
        ],
        "summary": "Get one work order",
        "operationId": "GetWorkOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderDetail"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Work orders"
        ],
        "summary": "Update editable work-order fields",
        "operationId": "UpdateWorkOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/work-orders/{id}/transitions": {
      "post": {
        "tags": [
          "Work orders"
        ],
        "summary": "Advance the guarded work-order lifecycle",
        "operationId": "TransitionWorkOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionWorkOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/work-orders/{id}/cancel": {
      "post": {
        "tags": [
          "Work orders"
        ],
        "summary": "Cancel a work order",
        "operationId": "CancelWorkOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelWorkOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderDetail"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CancelWorkOrderRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateWorkOrderRequest": {
        "type": "object",
        "properties": {
          "reference": {
            "type": "string",
            "nullable": true
          },
          "site": {
            "type": "string",
            "nullable": true
          },
          "summary": {
            "type": "string",
            "nullable": true
          },
          "priority": {
            "$ref": "#/components/schemas/WorkOrderPriority"
          },
          "dueAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "SortDirection": {
        "enum": [
          "Ascending",
          "Descending"
        ],
        "type": "string"
      },
      "TransitionWorkOrderRequest": {
        "type": "object",
        "properties": {
          "targetStatus": {
            "$ref": "#/components/schemas/WorkOrderStatus"
          },
          "assignee": {
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateWorkOrderRequest": {
        "type": "object",
        "properties": {
          "site": {
            "type": "string",
            "nullable": true
          },
          "summary": {
            "type": "string",
            "nullable": true
          },
          "priority": {
            "$ref": "#/components/schemas/WorkOrderPriority"
          },
          "dueAt": {
            "type": "string",
            "format": "date-time"
          },
          "assignee": {
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WorkOrderDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "reference": {
            "type": "string",
            "nullable": true
          },
          "site": {
            "type": "string",
            "nullable": true
          },
          "summary": {
            "type": "string",
            "nullable": true
          },
          "priority": {
            "$ref": "#/components/schemas/WorkOrderPriority"
          },
          "status": {
            "$ref": "#/components/schemas/WorkOrderStatus"
          },
          "dueAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "assignee": {
            "type": "string",
            "nullable": true
          },
          "cancellationReason": {
            "type": "string",
            "nullable": true
          },
          "isSlaRisk": {
            "type": "boolean"
          },
          "version": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WorkOrderListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "reference": {
            "type": "string",
            "nullable": true
          },
          "site": {
            "type": "string",
            "nullable": true
          },
          "summary": {
            "type": "string",
            "nullable": true
          },
          "priority": {
            "$ref": "#/components/schemas/WorkOrderPriority"
          },
          "status": {
            "$ref": "#/components/schemas/WorkOrderStatus"
          },
          "dueAt": {
            "type": "string",
            "format": "date-time"
          },
          "assignee": {
            "type": "string",
            "nullable": true
          },
          "isSlaRisk": {
            "type": "boolean"
          },
          "version": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WorkOrderListItemPagedResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkOrderListItem"
            },
            "nullable": true
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "WorkOrderPriority": {
        "enum": [
          "Standard",
          "Urgent",
          "Critical"
        ],
        "type": "string"
      },
      "WorkOrderSort": {
        "enum": [
          "DueAt",
          "CreatedAt",
          "UpdatedAt",
          "Reference",
          "Site",
          "Priority",
          "Status"
        ],
        "type": "string"
      },
      "WorkOrderStatus": {
        "enum": [
          "Planned",
          "Dispatched",
          "InProgress",
          "Blocked",
          "Completed",
          "Cancelled"
        ],
        "type": "string"
      }
    }
  }
}